Changed semantic representation of Symbols and types. Symbols are now, symbol -> typeid -> definition rather than symbol -> typeid -> symbol where the symbol is the definition and symbol at the same time. Adjusted design to align more with keywords.

+++ chern_core/src/id_types.rs
@@ -88,11 +88,11 @@ impl ExprId {
 
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
 pub struct ValueId {
-    pub id: usize,
+    pub id: u32,
 }
 
 impl ValueId {
-    pub fn new(id: usize) -> ValueId {
+    pub fn new(id: u32) -> ValueId {
         ValueId { id }
     }
 }
